rename liquidity test suite
