+++ programs/perpetuals/tests/native/tests_suite/mod.rs
@@ -1,9 +1,9 @@
-pub mod add_remove_liquidity;
 pub mod basic_interactions;
+pub mod liquidity;
 pub mod position;
 pub mod swap;
 
-pub use add_remove_liquidity::*;
 pub use basic_interactions::*;
+pub use liquidity::*;
 pub use position::*;
 pub use swap::*;
